server/requirements.txt
fastapi>=0.100
uvicorn[standard]>=0.20
python-dotenv>=1.0
